backend/requirements.txt
Flask
torch
diffusers
Pillow
accelerate
transformers
pandas
numpy
scikit-learn
faiss-cpu
